# src/main/resources/swagger/swaggerHeroi-Gen.yaml
swagger: "2.0"
info:
  description: "Este é o swagger dos microserviços /heroi /log e /ranking do Projeto de Bloco do Infnet, referente a avaliação do Professor Leonardo Silva da Gloria no semestre de 2021/2"
  version: "1.0.0"
  title: "Swagger Heroi-Gen"
host: "localhost:8091"
basePath: "/heroi-gen/"
tags:
- name: "heroi"
  description: ""
schemes:
- "https"
- "http"
paths:
  /herois:
    get:
      tags:
      - "heroi"
      summary: "Retorna a lista com o cadastro completo de heróis"
      description: "Retorna um array com todos os heróis cadastrados e seus atributos."
      operationId: "addHeroi"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      - "text/plain"
      responses:
        "200":
          description: "Operação realizada com sucesso"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Heroi"
        "404":
          description: "Nenhum herói cadastrado"
  /herois/{idHeroi}:
    get:
      tags:
      - "heroi"
      summary: "Retorna um herói e todos os seus atributos"
      description: "Retorna um héroi pelo id"
      operationId: "getHeroiById"
      produces:
      - "application/json"
      - "text/plain"
      parameters:
      - name: "idHeroi"
        in: "path"
        description: "ID do Herói a retornar"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Operação realizada com sucesso"
          schema:
            $ref: "#/definitions/Heroi"
        "400":
          description: "Id inválido"
        "404":
          description: "Herói não encontrado"
  /log/batalhas/{idBatalha}:
    get:
      tags:
      - "log"
      summary: "Retorna uma Batalha pelo id"
      description: "Retorna um array com todos os turnos de uma batalha em ordem cronológica"
      operationId: "getBatalhaById"
      produces:
        - "application/json"
        - "text/plain"
      parameters:
      - name: "idBatalha"
        in: "path"
        description: "ID da Batalha a retornar"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Operação realizada com sucesso"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Turno"
        "400":
          description: "Id inválido"
        "404":
          description: "Batalha não encontrada"
  /log/batalhas:
    post:
      tags:
      - "log"
      summary: "Salva turno de uma batalha"
      description: "Salva o turno de uma batalha"
      operationId: "saveTurnoBatalha"
      consumes:
      - "application/json"
      produces:
      - "text/plain"
      parameters:
      - in: "body"
        name: "Turno"
        description: "Turno da batalha a salvar"
        required: true
        schema:
          $ref: "#/definitions/Turno"
      responses:
        "200":
          description: "Operação realizada com sucesso"
        "400":
          description: "Formato de turno inválido"
        "404":
          description: "Batalha não localizada"
        "422":
          description: "Unprocessabled Entity"
  /ranking:
    get:
      tags:
      - "ranking"
      summary: "Retorna o ranking de pontuação"
      description: "Retorna um array com a pontuação final (em ordem decrescente) de todas as batalhas encerradas"
      operationId: "getRanking"
      produces:
        - "application/json"
        - "text/plain"
      responses:
        "200":
          description: "Operação realizada com sucesso"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Ranking"
        "404":
          description: "Nenhuma batalha foi encerrada"
  /ranking/{batalhaId}:
    post:
      tags:
      - "ranking"
      summary: "Salva uma batalha encerrada no ranking"
      description: "Salva uma batalha encerrada no ranking"
      operationId: "saveBatalha"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "text/plain"
      parameters:
      - in: "path"
        name: "batalhaId"
        description: "Id da Batalha a salvar"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Operação realizada com sucesso"
        "400":
          description: "id batalha inválido"
        "404":
          description: "Batalha não localizada"
definitions:
  Heroi:
    type: "object"
    required:
    - "heroi"
    - "pontosDeVida"
    - "forca"
    - "defesa"
    - "agilidade"
    - "qtdDadosDano"
    - "tamanhoDadosDano"
    properties:
      id:
        type: "integer"
        format: "int64"
      heroi:
        type: "string"
        example: "Guerreiro"
      pontosDeVida:
        type: "integer"
        format: "int64"
      forca:
        type: "integer"
        format: "int64"
      defesa:
        type: "integer"
        format: "int64"
      agilidade:
        type: "integer"
        format: "int64"
      qtdDadosDano:
        type: "integer"
        format: "int64"
      tamanhoDadosDano:
        type: "integer"
        format: "int64"
  Turno:
    type: "object"
    required:
    - "batalhaId"
    - "usuario"
    - "heroi"
    - "monstro"
    - "atacante"
    - "dano"
    properties:
      id:
        type: "integer"
        format: "int64"
      batalhaId:
        type: "integer"
        format: "int64"
      usuario:
        type: "string"
        example: "Zombie Slayer"
      heroi:
        type: "string"
        example: "Guerreiro"
      monstro:
        type: "string"
        example: "Morto Vivo"
      atacante:
        type: "string"
        description: "Entidade que está efetuando o ataque nesse turno"
        enum:
        - "HEROI"
        - "MONSTRO"
      dano:
        type: "integer"
        format: "int64"
  Ranking:
    type: "object"
    required:
    - "posicao"
    - "batalhaId"
    - "usuario"
    - "pontuacao"
    properties:
      id:
        type: "integer"
        format: "int64"
      posicao:
        type: "integer"
        format: "int64"
      batalhaId:
        type: "integer"
        format: "int64"
        example: "Guerreiro"
      usuario:
        type: "string"
        example: "Zombie Slayer"
      pontuacao:
        type: "integer"
        format: "int64"
externalDocs:
  description: "Veja o repositório do projeto no GitHub"
  url: "https://github.com/pgurjao/heroi-gen"